+++ src/main/java/com/verdantartifice/primalmagick/common/config/Config.java
@@ -18,6 +18,7 @@ import net.minecraftforge.fml.loading.FMLPaths;
  */
 @Mod.EventBusSubscriber
 public class Config {
+    protected static final String CATEGORY_WORLDGEN = "worldgen";
     protected static final String CATEGORY_MISC = "misc";
     
     protected static ForgeConfigSpec COMMON_CONFIG_SPEC;
@@ -26,6 +27,8 @@ public class Config {
     
     public static ForgeConfigSpec.BooleanValue SHOW_AFFINITIES;
     public static ForgeConfigSpec.BooleanValue SHOW_UNSCANNED_AFFINITIES;
+    public static ForgeConfigSpec.IntValue SHRINE_AVERAGE_DISTANCE_CHUNKS;
+    public static ForgeConfigSpec.IntValue SHRINE_MINIMUM_DISTANCE_CHUNKS;
     
     static {
         buildCommonConfigSpec();
@@ -35,6 +38,11 @@ public class Config {
     protected static void buildCommonConfigSpec() {
         // Define the common config file spec
         ForgeConfigSpec.Builder builder = new ForgeConfigSpec.Builder();
+        
+        builder.comment("Worldgen settings").push(CATEGORY_WORLDGEN);
+        SHRINE_AVERAGE_DISTANCE_CHUNKS = builder.comment("Average distance apart in chunks between shrine spawn attempts").defineInRange("shrineAverageDistanceChunks", 20, 1, 1000000);
+        SHRINE_MINIMUM_DISTANCE_CHUNKS = builder.comment("Minimum distance apart in chunks between shrine spawn attempts; must be less than average").defineInRange("shrineMinimumDistanceChunks", 10, 1, 1000000);
+        builder.pop();
         
         builder.comment("Misc settings").push(CATEGORY_MISC);
         SHOW_UNSCANNED_AFFINITIES = builder.comment("Show affinities of blocks and items even without scanning them").define("showUnscannedAffinities", false);