+++ src/main/java/com/verdantartifice/primalmagick/common/worldgen/features/StructureFeaturesPM.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.verdantartifice.primalmagick.PrimalMagick;
+import com.verdantartifice.primalmagick.common.config.Config;
 
 import net.minecraft.data.BuiltinRegistries;
 import net.minecraft.world.level.levelgen.StructureSettings;
@@ -31,19 +32,21 @@ public class StructureFeaturesPM {
     public static final RegistryObject<StructureFeature<ShrineConfig>> SHRINE = STRUCTURES.register("shrine", () -> new ShrineStructure(ShrineConfig.CODEC));
     
     public static void setupStructures() {
+        int shrineAvgDist = Config.SHRINE_AVERAGE_DISTANCE_CHUNKS.get();
+        int shrineMinDist = Math.min(Config.SHRINE_MINIMUM_DISTANCE_CHUNKS.get(), shrineAvgDist - 1);
         setupSpacingAndLand(
                 SHRINE.get(), 
                 new StructureFeatureConfiguration(
-                        20,         // Average distance apart in chunks between spawn attempts 
-                        10,         // Minimum distance apart in chunks between spawn attempts
-                        11893192),  // Seed modifier
+                        shrineAvgDist,  // Average distance apart in chunks between spawn attempts 
+                        shrineMinDist,  // Minimum distance apart in chunks between spawn attempts
+                        11893192),      // Seed modifier
                 true);
     }
     
     private static <S extends StructureFeature<?>> void setupSpacingAndLand(S structure, StructureFeatureConfiguration settings, boolean transformSurroundingLand) {
         StructureFeature.STRUCTURES_REGISTRY.put(structure.getRegistryName().toString(), structure);
         
-        // Modify surroudning land to conform to the bottom of the structure, if necessary
+        // Modify surrounding land to conform to the bottom of the structure, if necessary
         if (transformSurroundingLand) {
             StructureFeature.NOISE_AFFECTING_FEATURES = ImmutableList.<StructureFeature<?>>builder()
                     .addAll(StructureFeature.NOISE_AFFECTING_FEATURES)