+++ changelog.md
@@ -2,6 +2,7 @@ v2.0.2
 ------
 * Update model for ritual altar
 * Update texture for rock salt ore
+* Add config file entries to control shrine spacing in worldgen
 
 v2.0.1
 ------